Address new GCC build warning in new ssh related code
comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]

+++ src/sbbs3/answer.cpp
@@ -265,7 +265,7 @@ bool sbbs_t::answer()
 		}
 		l = 0;
 		if (cryptStatusOK(cryptGetAttributeString(ssh_session, CRYPT_SESSINFO_SSH_CHANNEL_TERMINAL, terminal, &l)) && l > 0) {
-			if (l < sizeof(terminal))
+			if (l < (int)sizeof(terminal))
 				terminal[l] = 0;
 			else
 				terminal[sizeof(terminal)-1] = 0;
